Rename folder to wat_compiler
