+++ src/wat_compiler/binary_writer.ts
