+++ src/wat_compiler/index.ts
