+++ src/wat_compiler/ir.ts
