+++ src/wat_compiler/lexer.ts
