+++ src/wat_compiler/parser.ts
