+++ test/ir.test.ts
@@ -1,4 +1,4 @@
-import { type Unfoldable } from '../src/wat2wasm/ir_types';
+import { type Unfoldable } from '../src/wat_compiler/ir_types';
 import { validTestCases as tc1 } from './resources/valid_function_bodies';
 
 test.each(tc1)('unfold simple_addition_sexpr', (testCase) => {
